feat: change uploading format from json to multitype/formdata; remove many IDs from request paths; improve response codes; change docs;

+++ apps/taysteer-backend/src/resources/users/user.service.ts
@@ -9,8 +9,6 @@ import {
   RateUserT,
   CheckAccessT,
   ValidateUserDataT,
-  ChangeUserImageT,
-  DeleteUserImageT,
   UserStringTypes,
 } from './user.service.types';
 import { UserT } from './user.types';
@@ -19,10 +17,10 @@ import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import bcrypt from 'bcryptjs';
-import { cloudinary } from 'configs/utils/cloudinary';
-import { PromiseController } from '../../utils/promise.controller';
-import { ADMIN_LOGIN } from 'configs/common/config';
+import { ADMIN_LOGIN, ADMIN_PASSWORD } from 'configs/common/config';
 import { UserRater } from './user.rater.model';
+import { deleteImage, uploadImage } from '../../utils/image.uploader';
+import { UserDataDto } from './user.dto';
 
 @Injectable()
 export class UsersService {
@@ -31,7 +29,10 @@ export class UsersService {
     private usersRepository: Repository<UserT>,
     @InjectRepository(UserRater)
     private userRatersRepository: Repository<UserRater>
-  ) {}
+  ) {
+    const admin = new User({ login: ADMIN_LOGIN, password: ADMIN_PASSWORD });
+    this.addUser(admin);
+  }
 
   checkAccess: CheckAccessT = async (
     user,
@@ -43,22 +44,30 @@ export class UsersService {
     return isOwner == shouldBeOwner || isAdmin;
   };
 
-  validateUserData: ValidateUserDataT = async (user) => {
+  validateUserData: ValidateUserDataT = async (userData, updating = false) => {
+    if (!updating) {
+      // If it's a data for creating a new user
+      if (!userData.login || !userData.password) return false; // Check if the login and password exists
+    } else {
+      // If it's a data for updating a user
+      if (
+        // Check if rating don't exist in the data
+        'rating' in userData ||
+        'ratings_count' in userData ||
+        'ratings_sum' in userData
+      )
+        return false;
+    }
     if (
-      !user.login ||
-      !user.password ||
-      (await this.usersRepository.findOne({ login: user.login })) || // Check if the same user already exists
-      user.image ||
-      user.rating ||
-      user.ratings_count ||
-      user.ratings_sum ||
-      user.name.length > 50 ||
-      user.login.length > 50 ||
-      user.password.length > 50 ||
-      user.description.length > 500
+      // Check if the data is valid
+      (await this.getByLogin(userData.login)) || // Check if the same user don't exists
+      userData.name.length > 50 ||
+      userData.login.length > 50 ||
+      userData.password.length > 50 ||
+      userData.description.length > 500
     )
       return false;
-    else return true;
+    return true;
   };
 
   getAllUsers: GetAllT = () => this.usersRepository.find();
@@ -68,8 +77,27 @@ export class UsersService {
   getByLogin: GetByLoginT = (login) =>
     this.usersRepository.findOne({ login: login });
 
-  addUser: AddUserT = async (user) => {
-    if (!(await this.validateUserData(user))) return false; // Validate data
+  addUser: AddUserT = async (userData, images) => {
+    if (!(await this.validateUserData(userData, false))) return false; // Validate data
+
+    const user = new User(userData); // Create user
+
+    // Upload image
+    if (images) {
+      // Get image
+      let image: NodeJS.ReadableStream;
+      for await (const img of images) {
+        image = img;
+        break;
+      }
+      // Upload image
+      const uploadedResponse = await uploadImage(
+        user.id,
+        image,
+        UserStringTypes.IMAGES_FOLDER
+      );
+      if (uploadedResponse) user.image = uploadedResponse;
+    }
     // Save user with password hash
     return this.usersRepository.save({
       ...user,
@@ -77,19 +105,37 @@ export class UsersService {
     });
   };
 
-  updateUser: UpdateUserT = async (id, newUser) => {
-    if (!(await this.validateUserData(newUser))) return false; // Validate data
+  updateUser: UpdateUserT = async (id, userData, images) => {
+    if (!(await this.validateUserData(userData, true))) return false; // Validate data
 
     const user = await this.getById(id);
     if (!user) return false; // Exit if user does not exist
+
+    // Upload image
+    if (images) {
+      // Get image
+      let image: NodeJS.ReadableStream;
+      for await (const img of images) {
+        image = img;
+        break;
+      }
+      // Upload image
+      const uploadedResponse = await uploadImage(
+        user.id,
+        image,
+        UserStringTypes.IMAGES_FOLDER
+      );
+      if (uploadedResponse) user.image = uploadedResponse;
+    }
+
     // Update the user
     return this.usersRepository.save({
       ...user,
       ...{
-        ...newUser,
+        ...userData,
         ...{
-          password: newUser.password // Set a hash of the password
-            ? await bcrypt.hash(newUser.password, bcrypt.genSaltSync(10))
+          password: userData.password // Set a hash of the password
+            ? await bcrypt.hash(userData.password, bcrypt.genSaltSync(10))
             : user.password,
         },
       },
@@ -101,8 +147,14 @@ export class UsersService {
     const user = await this.usersRepository.findOne(id, {
       relations: ['raters'],
     });
-    for (const rater of user.raters) this.userRatersRepository.delete(rater); // Delete the all raters
-    this.deleteUserImage(id); // Delete the image
+    if (!user) return false;
+    if (user.raters) {
+      // Delete the all raters
+      for (const rater of user.raters) {
+        this.userRatersRepository.delete(rater);
+      }
+    }
+    deleteImage(id, UserStringTypes.IMAGES_FOLDER); // Delete the image
     const deleteResult = await this.usersRepository.delete(id); // Delete the user
     return deleteResult.affected; // Return a result
   };
@@ -120,7 +172,7 @@ export class UsersService {
     // Validate data
     if (!rating) return false;
     else if (rating > 5) rating = 5;
-    
+
     // Get the user with relations
     const user = await this.usersRepository.findOne(id, {
       relations: ['raters'],
@@ -171,49 +223,5 @@ export class UsersService {
         raters: user.raters,
       },
     });
-  };
-
-  changeUserImage: ChangeUserImageT = async (id, fileReadStream) => {
-    const user = await this.getById(id); // Get a user
-    if (!user) return false; // Exit if user does not exist
-    if (user.image) await this.deleteUserImage(id); // Delete a old image
-    const promiseController = new PromiseController(); // Create a new promise controller
-    // Create upload stream
-    const uploadStream = cloudinary.uploader.upload_stream(
-      { public_id: user.id, folder: UserStringTypes.IMAGES_FOLDER },
-      (error, result) => {
-        if (!error) {
-          promiseController.resolve(result); // Return a response using the promise
-        } else promiseController.resolve(false);
-      }
-    );
-    fileReadStream.pipe(uploadStream); // Connect the file read stream
-    const uploadedResponse = await promiseController.promise; // Get a response
-    // Update the user
-    return uploadedResponse
-      ? this.usersRepository.save({
-          ...user,
-          ...{
-            image: uploadedResponse.url, // A new url
-          },
-        })
-      : false;
-  };
-
-  deleteUserImage: DeleteUserImageT = async (id) => {
-    const user = await this.getById(id); // Get user
-    if (!user) return false; // Exit if user does not exist
-    const response = await cloudinary.uploader.destroy(
-      `${UserStringTypes.IMAGES_FOLDER}/${user.id}`
-    ); // Delete a old image
-
-    return response.result == 'ok'
-      ? this.usersRepository.save({
-          ...user,
-          ...{
-            image: '', // Delete a url
-          },
-        })
-      : false;
   };
 }
